CLOUDIFY-1239 Added deploy goal to s3client. Moved src files to correct folder
