+++ cloudify/tools/maven-plugins/s3client-maven-plugin/pom.xml
@@ -3,7 +3,7 @@
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>org.cloudifysource</groupId>
 	<artifactId>s3client-maven-plugin</artifactId>
-	<version>1.1</version>
+	<version>1.2</version>
 	<packaging>maven-plugin</packaging>
 
 	<dependencies>
@@ -26,6 +26,12 @@
 			<version>1.2.1</version>
 		</dependency>
 
+		<dependency>
+			<groupId>org.apache.maven</groupId>
+			<artifactId>maven-project</artifactId>
+			<version>2.0.6</version>
+		</dependency>
+
 	</dependencies>
 
 	<distributionManagement>
@@ -34,5 +40,5 @@
 			<url>http://maven-repository.openspaces.org</url>
 		</repository>
 	</distributionManagement>
-	
+
 </project>
